Refactor intersect helpers into dedicated module
