+++ src/ECLlib/core.py
@@ -1,0 +1,10 @@
+"""Shared core components for ECLlib modules."""
+
+from .Files import DATA_file, ECL2IX_LOG, File, Restart
+
+__all__ = [
+    "File",
+    "DATA_file",
+    "Restart",
+    "ECL2IX_LOG",
+]
